Move App from component folder to container folder
